feat: remove temporary redirect to avoid duplicate default export in AdminManageUsersPage
